move KeyGen class to utils package
